# docs/reference/api-spec.yaml
...
openapi: 3.1.0
info:
  title: SceneFlow Backend
  version: 0.1.0
paths:
...
  /api/projects:
    get:
      tags:
      - projects
      summary: List Projects
      description: 'Every series, each carrying its current episode''s shots and a summary of the rest.


        Batched rather than serialized one project at a time: a list of series would otherwise

        run two queries per row. A project with no episode yet reads as empty here — creating

        one is a write, and a GET has no business doing that.'
      operationId: list_projects_api_projects_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Projects Api Projects Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - projects
      summary: Create Project
      operationId: create_project_api_projects_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Project Api Projects Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/projects/{project_id}/props:
    get:
      tags:
      - props
      summary: List Props
      operationId: list_props_api_projects__project_id__props_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Props Api Projects  Project Id  Props Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - props
      summary: Add Prop
      operationId: add_prop_api_projects__project_id__props_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePropRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Add Prop Api Projects  Project Id  Props Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}/props/{prop_id}:
    patch:
      tags:
      - props
      summary: Update Prop
      operationId: update_prop_api_projects__project_id__props__prop_id__patch
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: prop_id
        in: path
        required: true
        schema:
          type: string
          title: Prop Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePropRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Prop Api Projects  Project Id  Props  Prop Id  Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - props
      summary: Remove Prop
      operationId: remove_prop_api_projects__project_id__props__prop_id__delete
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: prop_id
        in: path
        required: true
        schema:
          type: string
          title: Prop Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}/props/{prop_id}/prompt:
    post:
      tags:
      - props
      summary: Draft Prop Prompt
      description: 'Queue the image prompt draft. Read it off the job result; it is never saved.


        Queued rather than awaited here because Starlette does not cancel a handler when the

        client disconnects, so a stop button could only ever hang up the browser while this kept

        running and billing. See `app/services/job_worker.py`.'
      operationId: draft_prop_prompt_api_projects__project_id__props__prop_id__prompt_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: prop_id
        in: path
        required: true
        schema:
          type: string
          title: Prop Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftPromptRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Draft Prop Prompt Api Projects  Project Id  Props  Prop Id  Prompt Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}/props/{prop_id}/image:
    post:
      tags:
      - props
      summary: Generate Prop Image
      description: Queue the prop's reference draw. The worker stores it and broadcasts `PROP_UPDATE`.
      operationId: generate_prop_image_api_projects__project_id__props__prop_id__image_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: prop_id
        in: path
        required: true
        schema:
          type: string
          title: Prop Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateReferenceImageRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Generate Prop Image Api Projects  Project Id  Props  Prop Id  Image Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - props
      summary: Upload Prop Image
      operationId: upload_prop_image_api_projects__project_id__props__prop_id__image_put
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: prop_id
        in: path
        required: true
        schema:
          type: string
          title: Prop Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadReferenceImageRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Upload Prop Image Api Projects  Project Id  Props  Prop Id  Image Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}/props/sheet:
    post:
      tags:
      - props
      summary: Merge Prop Sheet
      description: Tile every prop into one labelled sheet the storyboard render carries.
      operationId: merge_prop_sheet_api_projects__project_id__props_sheet_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Merge Prop Sheet Api Projects  Project Id  Props Sheet Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateProjectRequest:
      properties:
        title:
          type: string
          maxLength: 80
          title: Title
          default: ''
        description:
          type: string
          maxLength: 4000
          title: Description
          default: ''
        coverPrompt:
          type: string
          maxLength: 4000
          title: Coverprompt
          default: ''
        originalScript:
          type: string
          maxLength: 200000
          title: Originalscript
          default: ''
        productionSettings:
          $ref: '#/components/schemas/ProductionSettingsRequest'
      additionalProperties: false
      type: object
      title: CreateProjectRequest
    CreatePropRequest:
      properties:
        name:
          type: string
          maxLength: 80
          minLength: 1
          title: Name
        description:
          type: string
          maxLength: 4000
          title: Description
          default: ''
        ownerCharacterId:
          type: string
          maxLength: 64
          title: Ownercharacterid
          default: ''
        finalPrompt:
          type: string
          maxLength: 4000
          title: Finalprompt
          default: ''
        orderNum:
          type: integer
          maximum: 9999.0
          minimum: 0.0
          title: Ordernum
          default: 0
      additionalProperties: false
      type: object
      required:
      - name
      title: CreatePropRequest
...
    DraftPromptRequest:
      properties:
        name:
          type: string
          maxLength: 80
          title: Name
          default: ''
        description:
          type: string
          maxLength: 4000
          title: Description
          default: ''
        preset:
          type: string
          maxLength: 40
          title: Preset
          default: ''
        model:
          anyOf:
          - type: string
            maxLength: 160
          - type: 'null'
          title: Model
      additionalProperties: false
      type: object
      title: DraftPromptRequest
      description: 'Draft an image prompt for review.


        The fields are taken from the request rather than the stored row so the dialog can draft

        against edits the user has not saved yet. The instruction template is the built-in one

        in `prompt_service` and is not overridable — a second editable prompt field next to the

        one that actually draws only ever confused people about which was which.'
...
    GenerateReferenceImageRequest:
      properties:
        prompt:
          type: string
          maxLength: 4000
          title: Prompt
          default: ''
      additionalProperties: false
      type: object
      title: GenerateReferenceImageRequest
      description: Draw the reference from an approved prompt. Empty falls back to the stored one.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProductionSettingsRequest:
      properties:
        mode:
          anyOf:
          - type: string
            enum:
            - comic
            - drama
          - type: 'null'
          title: Mode
        aspectRatio:
          anyOf:
          - type: string
            enum:
            - '9:16'
            - '16:9'
            - '1:1'
          - type: 'null'
          title: Aspectratio
        width:
          anyOf:
          - type: integer
            maximum: 4096.0
            minimum: 256.0
          - type: 'null'
          title: Width
        height:
          anyOf:
          - type: integer
            maximum: 4096.0
            minimum: 256.0
          - type: 'null'
          title: Height
        fps:
          anyOf:
          - type: integer
            enum:
            - 24
            - 30
          - type: 'null'
          title: Fps
        targetDurationMs:
          anyOf:
          - type: integer
            maximum: 600000.0
            minimum: 10000.0
          - type: 'null'
          title: Targetdurationms
        language:
          anyOf:
          - type: string
            maxLength: 20
            minLength: 1
          - type: 'null'
          title: Language
        stylePrompt:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Styleprompt
        negativePrompt:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Negativeprompt
        currentStage:
          anyOf:
          - type: string
            enum:
            - script
            - bible
            - storyboard
            - audio
            - timeline
            - export
          - type: 'null'
          title: Currentstage
      additionalProperties: false
      type: object
      title: ProductionSettingsRequest
      description: 'Every field optional so a PATCH can carry just the one the user changed.


        Range checks live here for the OpenAPI schema; `project_service.production_settings`

        stays the single place that resolves defaults and cross-field consistency.'
...
    UpdatePropRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 80
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Description
        ownerCharacterId:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Ownercharacterid
        finalPrompt:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Finalprompt
        orderNum:
          anyOf:
          - type: integer
            maximum: 9999.0
            minimum: 0.0
          - type: 'null'
          title: Ordernum
      additionalProperties: false
      type: object
      title: UpdatePropRequest
...
    UploadReferenceImageRequest:
      properties:
        imageData:
          type: string
          maxLength: 14000000
          minLength: 1
          title: Imagedata
      additionalProperties: false
      type: object
      required:
      - imageData
      title: UploadReferenceImageRequest
      description: A reference the user drew themselves. `data:image/...;base64,` — never multipart.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError